Fuctional Shop. Dynamically find shop UI elements and cache them for use while opening shop. Implemented player levels to set items in shop. Super values are randomly scaled based on player level.

+++ Assets/Scripts/Enemies/Controller.cs
@@ -7,7 +7,7 @@ namespace Enemies
 {
     public partial class Controller : MonoBehaviour
     {
-        public static event System.Action<AttackType, float, float, Vector2> OnEnemyDeath;
+        public static event System.Action<AttackType, int, float, Vector2> OnEnemyDeath;
 
         [FoldoutGroup("Movement Settings"), SerializeField, Tooltip("The movement speed of the enemy")]
         private float movementSpeed = 3.0f;
@@ -18,7 +18,7 @@ namespace Enemies
         [FoldoutGroup("Damage Settings"), SerializeField, Tooltip("The time in seconds between damage ticks")]
         private float damageInterval = 1.0f;
         [FoldoutGroup("Death Values"), SerializeField, Tooltip("The value of the enemy when it dies")]
-        private float deathValue = 1f;
+        private int deathValue = 1;
         [FoldoutGroup("Death Values"), SerializeField, Tooltip("The chance (0 to 1) to spawn an item on enemy death"), Range(0f, 1f)]
         private float itemSpawnChance = 0f;
